Schema user_products altered

+++ database/migrations/2022_01_18_141500_create_user_products_table.php
@@ -15,8 +15,8 @@ class CreateUserProductsTable extends Migration
     {
         Schema::create('user_products', function (Blueprint $table) {
             $table->id();
-            $table->integer('user_id')->unsigned();
-            $table->integer('product_id')->unsigned();
+            $table->bigInteger('user_id')->unsigned();
+            $table->bigInteger('product_id')->unsigned();
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
             $table->integer('amount');
